+++ packages/core-mobile/app/new/routes/(signedIn)/(modals)/accountSettings/addressBook/addContact.tsx
@@ -1,3 +1,3 @@
-import { AddContactScreen } from 'features/accountSettings/components/addressBook/AddContactScreen'
+import { AddContactScreen } from 'features/accountSettings/screens/addressBook/AddContactScreen'
 
 export { AddContactScreen as default }
