+++ packages/core-mobile/app/new/routes/(signedIn)/(modals)/accountSettings/addressBook/index.tsx
@@ -1,3 +1,3 @@
-import { AddressBookScreen } from 'features/accountSettings/components/addressBook/AddressBookScreen'
+import { AddressBookScreen } from 'features/accountSettings/screens/addressBook/AddressBookScreen'
 
 export { AddressBookScreen as default }
